# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "xg backend rest api",
        "title": "xg REST API",
        "termsOfService": "https://localhost:8088/v1",
        "contact": {},
        "version": "0.1 alpha"
    },
    "paths": {
...
        "/api/order": {
            "post": {
                "description": "create an order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "createOrder",
                "parameters": [
                    {
                        "description": "create request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.IdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/order/{id}": {
            "get": {
                "description": "get order by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "getOrderByID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.OrderInfoWithRecords"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/order/{id}/mark": {
            "post": {
                "description": "add mark to order by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "addOrderMark",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "create mark",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.OrderMarkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/order/{id}/pay": {
            "put": {
                "description": "pay order by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "payOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "order pay request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.OrderPayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/order/{id}/payback": {
            "put": {
                "description": "payback order by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "paybackOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "order payback request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.OrderPayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/order/{id}/revoke": {
            "put": {
                "description": "revoke order by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "revokeOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/order/{id}/signup": {
            "put": {
                "description": "signup order by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "signupOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "order id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "order signup request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.OrderPayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
...
        "/api/orders": {
            "get": {
                "description": "search order with condition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "searchOrder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search order with student_ids",
                        "name": "student_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search order with to_org_ids",
                        "name": "to_org_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search order with intent_subjects",
                        "name": "intent_subjects",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "search order with publisher_id",
                        "name": "publisher_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search order with status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search order order by column name",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "order list page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "order list page index",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.OrderInfoList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
        "/api/orders/author": {
            "get": {
                "description": "search order in author with condition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "searchOrderWithAuthor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "search order with student_ids",
                        "name": "student_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search order with to_org_ids",
                        "name": "to_org_ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search order with intent_subjects",
                        "name": "intent_subjects",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search order with status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search order order by column name",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "order list page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "order list page index",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.OrderInfoList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
...
        "/api/payment/{id}/review/accept": {
            "put": {
                "description": "accept payment by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "order"
                ],
                "summary": "acceptPayment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "payment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/route.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "intent_subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "student_id": {
                    "type": "integer"
                },
                "to_org_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.OrderInfoWithRecords": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "intent_subject": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "org_name": {
                    "type": "string"
                },
                "paymentInfo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.OrderPayRecord"
                    }
                },
                "publisher_id": {
                    "type": "integer"
                },
                "publisher_name": {
                    "type": "string"
                },
                "remarkInfo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.OrderRemarkRecord"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "student_id": {
                    "type": "integer"
                },
                "student_summary": {
                    "$ref": "#/definitions/entity.StudentSummaryInfo"
                },
                "to_org_id": {
                    "type": "integer"
                }
            }
        },
        "entity.OrderMarkRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "order_id": {
                    "type": "integer"
                }
            }
        },
...
        "route.IdResponse": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "route.Response": {
            "type": "object",
            "properties": {
                "err_msg": {
                    "type": "string"
                }
            }
        },
...
    }
}